last clean

+++ src/main/java/martinez/ruben/api/entity/User.java
@@ -28,7 +28,7 @@ public class User {
     private String email;
 
     private Integer maxWave;
-    private Integer playtime;
+    private float playtime;
     private Integer games;
     private Integer kills;
     private Integer maxKills;
